requirements.txt
streamlit>=1.24.0
pytest>=7.0.0
langchain 
langchain-groq 
groq
python-dotenv
langchain-community
# llama-cpp-python --prefer-binary
pydocstyle
radon
pytest-json-report


# pip install -r requirements.txt

# pytest --json-report --json-report-file=storage/reports/pytest_results.json
